Delegated the JDBC stuff strictly to the JDBC-related packages and artifacts.
Removed the concept of "data type" and instead will allow the higher layers to deal with that concept, as it's too much complication to deal with it at the lowest levels. In particular, all data is now stored as strings, so there will have to be an encoder/decoder layer implemented on top of the object store.