As I was saying... :D
Object locks now require two steps: one is the history-level lock, which
is re-entrant, and one is the object-level lock, which is not.  This is
important because it allows us to eliminate some weird race conditions
that can be triggered when multiple threads attempt to export objects
from the same history (i.e. when Documentum export predicates include
"(all)").

However, we take care to ensure that the locks remain atomic in order to
preserve both referential integrity, and responsiveness.