Modified the filename fixing to use history ID vs. object ID, which is much more in line with what the filename mapper actually generates now.
The mappings it used to generate were based on the ObjectID, but that
makes little sense since a) it's scanning over only the CURRENT versions
and generating mappings based on that object ID, and b) the true value
of this mapping table is for when we're moving from Documentum to an ECM
that doesn't support per-version naming schemes, and has no filename
collision issues (i.e. allows duplicate filenames).