Modified how ExportTargets are built to avoid extraction failures
Extractions could fail if the search thread found an object it didn't
know what to make of. This was previously left up to the engines to
decide whether to explode or return a null ExportTarget which would then
be ignored by the scanner pipeline.

This behavior is incorrect since an explosion will cause the extraction
to stop short (semi-)quietly. Instead, what should happen is that the
unrecognized object is identified, logged, but then ignored (by way of
producing a null ExportTarget).

Unfortunately, this has a cascade effect unto other methods which also
required adjustment for the new intended behavior.